+++ tests/regression/struct.py
@@ -340,7 +340,46 @@ class TestTS(StructMixin, test.TestCase):
         self.assertEqual(ts.get(date(1990,1,1)),None)
         self.assertEqual(ts.get(date(1990,1,1),1),1)
         self.assertRaises(KeyError, lambda : ts[date(1990,1,1)])
-    
+        
+    def testPop(self):
+        session = self.session()
+        with session.begin():
+            ts = session.add(odm.TS())
+            ts.update(zip(dates,values))
+        dt = dates[5]
+        self.assertTrue(dt in ts)
+        v = ts.pop(dt)
+        self.assertTrue(v)
+        self.assertFalse(dt in ts)
+        self.assertRaises(KeyError, ts.pop, dt)
+        self.assertEqual(ts.pop(dt,'bla'), 'bla')
+        
+    def test_rank_ipop(self):
+        session = self.session()
+        with session.begin():
+            ts = session.add(odm.TS())
+            ts.update(zip(dates,values))
+        dt = dates[5]
+        value = ts.get(dt)
+        r = ts.rank(dt)
+        all_dates = list((d.date() for d in ts.itimes()))
+        self.assertEqual(all_dates[r], dt)
+        value2 = ts.ipop(r)
+        self.assertEqual(value, value2)
+        self.assertFalse(dt in ts)
+        
+    def test_pop_range(self):
+        session = self.session()
+        with session.begin():
+            ts = session.add(odm.TS())
+            ts.update(zip(dates,values))
+        all_dates = list((d.date() for d in ts.itimes()))
+        range = list(ts.range(all_dates[5],all_dates[15]))
+        self.assertTrue(range)
+        range2 = list(ts.pop_range(all_dates[5],all_dates[15]))
+        self.assertEqual(range, range2)
+        for dt,_ in range:
+            self.assertFalse(dt in ts) 
 
 class TestString(StructMixin, test.TestCase):
     structure = odm.String